% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_analysis_landscapes.R
\name{plot_analysis_landscapes}
\alias{plot_analysis_landscapes}
\title{Plot the landscapes used for simulation and the classified used for analysis}
\usage{
plot_analysis_landscapes(movementDataAll)
}
\value{
a
}
\description{
A
}
