% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read in all snake data and spatial data}
\usage{
read_data(species, hypothesis, landscape)
}
\arguments{
\item{species}{c("OPHA", "PYBI", "BUCA", "BUFA")}

\item{landscape}{c("binary", "continuous")}
}
\value{
a
}
\description{
A
}
