% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rmd.R
\name{render_rmd}
\alias{render_rmd}
\title{Render the manuscript rmd}
\usage{
render_rmd(
  fileIN,
  fileOUT,
  modelExtracts = NA,
  effectPlots = NA,
  allSpecCurve = NA,
  optionsCompleteList = NA
)
}
\arguments{
\item{allSpecCurve}{ggplot objects for plots showing the spec curves}

\item{modelExtractsTarget}{table of R2 and effects for quick reference in the rmd}

\item{effectPlotsTarget}{ggplot objects for plots showing the effects for quick reference in the rmd}
}
\value{
Nothing, PDF (or output) will be saved to a folder.
}
\description{
A
}
